requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
coverage
django-cors-headers